Fix truncation from double to float

+++ code/q3_ui/ui_players.c
@@ -739,7 +739,7 @@ void UI_DrawPlayer( float x, float y, float w, float h, playerInfo_t *pi, int ti
 	refdef.fov_x = (int)((float)refdef.width / 640.0f * 90.0f);
 	xx = refdef.width / tan( refdef.fov_x / 360 * M_PI );
 	refdef.fov_y = atan2( refdef.height, xx );
-	refdef.fov_y *= ( 360 / M_PI );
+	refdef.fov_y *= (float)( 360 / M_PI );
 
 	// calculate distance so the player nearly fills the box
 	len = 0.7 * ( maxs[2] - mins[2] );
